Add files via upload
Programa modificado con los cambios hechos al algoritmo.
para la entrada:
TEXT: Escherichia coli is able to use trehalose [O-U-D-glucosyl-a-D-glucoside] as a carbon source but can also synthesize and accumulate trehalose as an osmoprotectant when grown in high-osmolarity medium.
Da la salida:
Escherichia coli is able to use trehalose [O-U-D-glucosyl-a-D-glucoside] as a carbon source when grown in high-osmolarity medium. 
Escherichia coli is able to can also synthesize and accumulate trehalose as an osmoprotectant when grown in high-osmolarity medium. 
Escherichia coli is able to use trehalose [O-U-D-glucosyl-a-D-glucoside] as a carbon source but can also synthesize trehalose as an osmoprotectant when grown in high-osmolarity medium. 
Escherichia coli is able to use trehalose [O-U-D-glucosyl-a-D-glucoside] as a carbon source but can also accumulate trehalose as an osmoprotectant when grown in high-osmolarity medium. 
Debe dar la salida:
Escherichia coli is able to use trehalose [